🔧 Unsloth dtype 호환성 문제 해결
- trainer.py: Unsloth dtype을 float32에서 bfloat16/float16으로 수정
- bf16=true일 때 torch.bfloat16, 아니면 torch.float16 사용
- Unsloth AssertionError 해결

이제 Unsloth가 정상 동작할 수 있음

+++ code/trainer.py
@@ -1068,7 +1068,8 @@ class DialogueSummarizationTrainer:
                 model_name=model_checkpoint,
                 max_seq_length=self.config["tokenizer"].get("encoder_max_len", 512)
                 + self.config["tokenizer"].get("decoder_max_len", 200),
-                dtype=torch.float16 if self.config["training"].get("fp16") else torch.float32,
+                # 🔧 Unsloth dtype 호환성 수정: float16 또는 bfloat16만 지원
+                dtype=torch.bfloat16 if self.config["training"].get("bf16") else torch.float16,
                 load_in_4bit=qlora_config.get("load_in_4bit", True),
             )
